refactor: modularize converter code, update docstrings
